refactor(entities): extract <ReviewActions> to shared component
ECR-03B PR 1 of 3 — pure refactor with no behaviour change. Moves the
resolve UI out of routes/review-entities/index.tsx into a shared
components/review-actions.tsx so the same affordances can render inline
on the entities tab in PR 2 of this series.

The page's existing integration tests continue to cover end-to-end
behaviour (they mount the page which mounts the extracted component);
adds a small isolated test file pinning the component's own contract
plus the new optional onResolved callback that PR 2 will consume.

Exports the cache key constants (LIST_KEY, COUNT_KEY, detailKey) from
the new module so the page and the inline host share one source of
truth — keeps optimistic-removal invalidations correct across both
surfaces.
